chore: Switch back to the original idea since using context + CellRendererComponent made things more complicated
